requirements.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==6.1.0  # https://github.com/un33k/python-slugify
Pillow==9.0.1  # https://github.com/python-pillow/Pillow
rcssmin==1.1.0  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.0.0  # https://github.com/evansd/whitenoise
redis==4.1.4  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==4.0.2  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.49.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-phonenumber-field==6.0.0  # https://github.com/stefanfoulis/django-phonenumber-field
phonenumbers==8.12.45
django-extensions==3.1.5  # https://github.com/django-extensions/django-extensions
django-jalali>=5.1.0
django-localflavor==3.1
django-multiselectfield==0.1.12
ippanel==1.1.0

gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
psycopg2==2.9.3  # https://github.com/psycopg/psycopg2

# Django
# ------------------------------------------------------------------------------
django-anymail==8.5  # https://github.com/anymail/django-anymail
django-debug-toolbar==3.2.4  # https://github.com/jazzband/django-debug-toolbar